Fix "not using used parts in optimized search"
It seems that currently, when using --include_used=all (or any other used
parts), the fact that I want to include used parts is not actually used in the
optimizer. The reason is that while OptimizerBase->Load() takes an
'allowed_used' argument, and passes it further down, that information is
actually not given to it when it is called. This patch fixes this.

Fixes issue 2.